Move PoolID to separate header, adjust pool header includes
