+++ src/core/pool_type.hpp
@@ -11,6 +11,7 @@
 #define POOL_TYPE_HPP
 
 #include "enum_type.hpp"
+#include "pool_id_type.hpp"
 #include "../debug_dbg_assert.h"
 #include <vector>
 
@@ -25,63 +26,6 @@ using PoolTypes = EnumBitSet<PoolType, uint8_t>;
 static constexpr PoolTypes PT_ALL = {PoolType::Normal, PoolType::NetworkClient, PoolType::NetworkAdmin, PoolType::Data};
 
 typedef std::vector<struct PoolBase *> PoolVector; ///< Vector of pointers to PoolBase
-
-/** Non-templated base for #PoolID for use with type trait queries. */
-struct PoolIDBase {};
-
-/**
- * Templated helper to make a PoolID a single POD value.
- *
- * Example usage:
- *
- *   using MyType = PoolID<int, struct MyTypeTag, 16, 0xFF>;
- *
- * @tparam TBaseType Type of the derived class (i.e. the concrete usage of this class).
- * @tparam TTag An unique struct to keep types of the same TBaseType distinct.
- * @tparam TEnd The PoolID at the end of the pool (equivalent to size).
- * @tparam TInvalid The PoolID denoting an invalid value.
- */
-template <typename TBaseType, typename TTag, TBaseType TEnd, TBaseType TInvalid>
-struct EMPTY_BASES PoolID : PoolIDBase {
-	static inline constexpr bool fmt_as_base = true;
-	static inline constexpr bool serialisation_as_base = true;
-	static inline constexpr bool saveload_primitive_type = true;
-	static inline constexpr bool string_parameter_as_base = true;
-	static inline constexpr bool script_stack_value_as_base = true;
-	static inline constexpr bool integer_type_hint = true;
-	static inline constexpr bool hash_as_base = true;
-
-	using BaseType = TBaseType;
-
-	constexpr PoolID() = default;
-	constexpr PoolID(const PoolID &) = default;
-	constexpr PoolID(PoolID &&) = default;
-
-	explicit constexpr PoolID(const TBaseType &value) : value(value) {}
-
-	constexpr PoolID &operator =(const PoolID &rhs) { this->value = rhs.value; return *this; }
-	constexpr PoolID &operator =(PoolID &&rhs) { this->value = std::move(rhs.value); return *this; }
-
-	/* Only allow conversion to BaseType via method. */
-	constexpr TBaseType base() const noexcept { return this->value; }
-	constexpr TBaseType &edit_base() { return this->value; }
-
-	static constexpr PoolID Begin() { return PoolID{}; }
-	static constexpr PoolID End() { return PoolID{static_cast<TBaseType>(TEnd)}; }
-	static constexpr PoolID Invalid() { return PoolID{static_cast<TBaseType>(TInvalid)}; }
-
-	constexpr auto operator++() { ++this->value; return this; }
-	constexpr auto operator+(const std::integral auto &val) const { return this->value + val; }
-
-	constexpr bool operator==(const PoolID<TBaseType, TTag, TEnd, TInvalid> &rhs) const { return this->value == rhs.value; }
-	constexpr auto operator<=>(const PoolID<TBaseType, TTag, TEnd, TInvalid> &rhs) const { return this->value <=> rhs.value; }
-
-	constexpr bool operator==(const size_t &rhs) const { return this->value == rhs; }
-	constexpr auto operator<=>(const size_t &rhs) const { return this->value <=> rhs; }
-private:
-	/* Do not explicitly initialize. */
-	TBaseType value;
-};
 
 /** Base class for base of all pools. */
 struct PoolBase {
