+++ src/economy_type.h
@@ -12,7 +12,7 @@
 
 #include "core/overflowsafe_type.hpp"
 #include "core/enum_type.hpp"
-#include "core/pool_type.hpp"
+#include "core/pool_id_type.hpp"
 #include <array>
 
 typedef OverflowSafeInt64 Money;
