+++ src/network/core/tcp_listen.h
@@ -14,7 +14,6 @@
 
 #include "tcp.h"
 #include "../network.h"
-#include "../../core/pool_type.hpp"
 #include "../../debug.h"
 #include "table/strings.h"
