+++ src/signs_type.h
@@ -10,7 +10,7 @@
 #ifndef SIGNS_TYPE_H
 #define SIGNS_TYPE_H
 
-#include "core/pool_type.hpp"
+#include "core/pool_id_type.hpp"
 
 /** The type of the IDs of signs. */
 using SignID = PoolID<uint16_t, struct SignIDTag, 64000, 0xFFFF>;
